fix about section
