+++ src/components/ActionButton/ActionButton.module.scss
@@ -41,7 +41,6 @@
   &.tertiary {
     background-color: $primaryColor;
     color: $backgroundColor;
-    padding: 27px;
 
     &:hover {
       background-color: rgba($primaryColor, 0.95);
