--- src/components/BackgroundTriangle/BackgroundTriangle.module.scss
@@ -1,23 +1,0 @@
-@import "../../styles/styles";
-
-.backgroundTriangle {
-  position: absolute;
-  z-index: auto;
-  top: 0;
-  left: 0;
-  width: 100%;
-}
-
-.triangle {
-  position: absolute;
-  width: 0;
-  height: 0;
-}
-
-.dottedTriangle {
-  background-image: radial-gradient($primaryColor 20%, transparent 0);
-  background-size: 10px 10px;
-  width: 100%;
-  height: 100%;
-}
-
