--- src/components/BackgroundTriangle/index.ts
@@ -1,2 +1,0 @@
-import BackgroundTriangle from "./BackgroundTriangle";
-export default BackgroundTriangle;
