--- src/components/BackgroundTriangle/types.ts
@@ -1,5 +1,0 @@
-export interface Props {
-  isOdd: boolean;
-  showLightBlueTriangle?: boolean;
-  backgroundColor: string;
-}
