+++ src/components/CtaButtons/index.ts
@@ -1,0 +1,2 @@
+import CtaButtons from "./CtaButtons";
+export default CtaButtons;
