+++ src/components/CtaButtons/types.ts
@@ -1,0 +1,5 @@
+export interface Props {
+  className?: string;
+}
+
+export interface State {}
