+++ src/components/History/index.ts
@@ -1,0 +1,2 @@
+import History from "./History";
+export default History;
