+++ src/components/InfoIcon/InfoIcon.module.scss
@@ -5,7 +5,7 @@
   align-items: center;
   color: $backgroundColor;
   transition: color 0.25s ease-in-out 0s;
-  margin: 0 75px 25px 0;
+  margin: 0 25px;
   cursor: default;
 
   text-transform: none;
