+++ src/components/Speakers/Speakers.module.scss
@@ -4,8 +4,8 @@
   @extend %banner;
 }
 
-.subtitle {
-  @extend %subtitle;
+.tag {
+  @extend %tag;
 }
 .title {
   @extend %title;
