+++ src/components/Speakers/Speakers.tsx
@@ -17,19 +17,9 @@ export default class Speakers extends React.PureComponent<Props, State> {
 
   render() {
     const { className, speakers, type } = this.props;
-    const cssClasses = classNames(styles.speakers, className);
-    const Resources = resourcesService.getResources();
-    const actionButtonType = type === "odd" ? "tertiary" : "secondary";
 
     return (
       <div>
-      <div className={styles.banner}>
-        <span className={styles.subtitle}>{Resources.pages.speakers}</span>
-        <span className={styles.title}>
-        {Resources.titles.speakersPage}
-        </span>
-        <ActionButton type={actionButtonType} text={Resources.buttons.wantToBeSpeaker} url={constants.speakerCallUrl} />
-        </div>
         <Team className={styles.speakersList} team={speakers} type={type} />
       </div>
     );
