+++ src/components/SponsorshipPackages/SponsorshipPackages.module.scss
@@ -9,8 +9,8 @@
   @extend %banner;
 }
 
-.subtitle {
-  @extend %subtitle;
+.tag {
+  @extend %tag;
 }
 
 .title {
