+++ src/pages/ConductPage/ConductPage.module.scss
@@ -14,7 +14,7 @@
   font-size: 21px;
   column-count: 2;
   margin: 0 35px;
-  padding: 0 100px;
+  padding: 0 100px 100px;
 }
 
 @media
