+++ src/pages/ExecutiveTeamPage/ExecutiveTeamPage.module.scss
@@ -5,8 +5,8 @@
   height: 400px;
 }
 
-.subtitle {
-  @extend %subtitle;
+.tag {
+  @extend %tag;
 }
 .title {
   @extend %title;
