+++ src/pages/SchedulePage/SchedulePage.tsx
@@ -1,8 +1,10 @@
 import React from "react";
+import { resourcesService } from "../../services";
 import styles from "./SchedulePage.module.scss";
 
 export default class SchedulePage extends React.PureComponent {
   render() {
-    return <div className={styles.schedulePage}>Próximamente</div>;
+    const Resources = resourcesService.getResources();
+    return <div className={styles.schedulePage}>{Resources.buttons.soon}</div>;
   }
 }
