+++ src/pages/SpeakersPage/SpeakersPage.module.scss
@@ -7,3 +7,15 @@
   justify-content: center;
   flex: 1;
 }
+
+.banner {
+  @extend %banner;
+}
+
+.tag {
+  @extend %tag;
+}
+.title {
+  @extend %title;
+}
+
